Перемещение классов в новые пакеты
