--- app/src/main/java/com/practicum/playlistmaker/TrackAdapter.kt
@@ -1,28 +1,0 @@
-package com.practicum.playlistmaker
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-import com.practicum.playlistmaker.domain.model.Track
-
-class TrackAdapter(
-    private val tracks: List<Track>,
-    private val click: (Track) -> Unit
-) : RecyclerView.Adapter<TrackViewHolder>(){
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TrackViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.item_track, parent, false)
-        return TrackViewHolder(view)
-    }
-
-    override fun getItemCount(): Int {
-        return tracks.size
-    }
-
-    override fun onBindViewHolder(holder: TrackViewHolder, position: Int) {
-        holder.bind(tracks[position])
-        holder.itemView.setOnClickListener{
-            click(tracks[position])
-        }
-    }
-}
