+++ app/src/main/java/com/practicum/playlistmaker/ui/LibraryActivity.kt
@@ -1,10 +1,11 @@
-package com.practicum.playlistmaker
+package com.practicum.playlistmaker.ui
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.practicum.playlistmaker.R
 
 class LibraryActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
