+++ app/src/main/java/com/practicum/playlistmaker/ui/PlayerActivity.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker
+package com.practicum.playlistmaker.ui
 
 import android.icu.text.SimpleDateFormat
 import android.media.MediaPlayer
@@ -13,8 +13,9 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+import com.practicum.playlistmaker.R
 import com.practicum.playlistmaker.domain.model.Track
-import com.practicum.playlistmaker.ui.SearchActivity
+import com.practicum.playlistmaker.ui.search.SearchActivity
 import java.util.Locale
 
 class PlayerActivity : AppCompatActivity() {
