+++ app/src/main/res/layout/activity_search.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main_search"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.SearchActivity"
+    tools:context=".ui.search.SearchActivity"
     android:orientation="vertical">
 
     <LinearLayout
